Rewrite large portions of library
Previously, this library could be thought of as a general abstraction of
the PostgreSQL wire protocol, instead of specifically as a client for the
wire protocol.  There were large amounts of code not used, because they
could theoretically be used for the server-side of the protocol, instead
of only the client side.

This rewrites large portions of the library and strips out what isn't
needed.  It fixes issues like transaction status always being unknown
(which was due to Ruby 1.8 dependent code).

This combine the binary_reader, binary_writer, buffer, and byteorder code
into a single postgres-pr/buffer file.

This makes the underlying PostgresPR::Connection and PostgresPR::Result
objects compatible with what Sequel expects.  This removes the
postgres-pr/postgres-compat file, since some changes will not be
compatible with older versions of Sequel, and I want to avoid older
versions of Sequel picking this newer adapter up. A new sequel/postgres-pr
file is added, which is what an updated version of Sequel will look for.