Add test suite with 100% line/branch coverage
Use Sequel's postgres adapter tests to handle most of the testing.  What
isn't covered by that, cover in a separate test file.  Sequel should be
checked out to ../sequel when running the test_cov task.

Some parts of the library require specific configuration on the PostgreSQL
server to enable different auth methods.  The test/connection_recorder.rb
exists to make it easy to record socket traffic between the client and
the server.  Recorded socket traffic is used for the tests.